Split up testrunner setup utils
The setup utils file was getting close to 1000 lines and becoming hard
to navigate.
